organized application ids of experiments
