+++ projects/CalculatorExperiments/app/build.gradle
@@ -8,7 +8,7 @@ android {
     buildToolsVersion "30.0.2"
 
     defaultConfig {
-        applicationId 'android.support.constraint.calc'
+        applicationId 'androidx.constraintlayout.experiments.calculator'
         minSdkVersion 16
         targetSdkVersion 30
         versionCode 1
