+++ projects/ConstraintLayoutValidation/app/src/androidTest/java/androidx/constraintlayout/validation/SampleTest.kt
@@ -1,4 +1,4 @@
-package com.example.developertemplate
+package androidx.constraintlayout.validation
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -14,11 +14,12 @@ import org.junit.Assert.*
  * See [testing documentation](http://d.android.com/tools/testing).
  */
 @RunWith(AndroidJUnit4::class)
-class ExampleInstrumentedTest {
+class SampleTest {
+
     @Test
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.example.developertemplate", appContext.packageName)
+        assertEquals("androidx.constraintlayout.validation.constraintlayout.widget.test", appContext.packageName)
     }
 }
