+++ projects/FoldableExperiments/app/build.gradle
@@ -8,7 +8,7 @@ android {
     buildToolsVersion "30.0.2"
 
     defaultConfig {
-        applicationId "com.example.developertemplate"
+        applicationId "androidx.constraintlayout.experiments.foldable"
         minSdkVersion 16
         targetSdkVersion 30
         versionCode 1
